+++ projects/ngx-qgrid/src/lib/column-list/column-list.service.ts
@@ -4,48 +4,41 @@ import { parseFactory } from '@qgrid/core/services/convert';
 import { ColumnListHost } from '@qgrid/core/column-list/column.list.host';
 import { ColumnModel } from '@qgrid/core/column-type/column.model';
 import { GridPlugin } from '../plugin/grid-plugin';
+import { Lazy } from '@qgrid/core/infrastructure/lazy';
 
 @Injectable()
 export class ColumnListService {
-	private host: ColumnListHost;
+	private host = new Lazy(() => {
+		const canCopy = (key: string, source, target) =>
+			target.hasOwnProperty(key) && !isUndefined(source[key]);
+
+		return new ColumnListHost(this.plugin.model, canCopy, parseFactory);
+	});
 
 	constructor(private plugin: GridPlugin) {
 	}
 
 	add(column: ColumnModel) {
-		this.view.add(column);
+		this.host.instance.add(column);
 	}
 
 	copy(target, source) {
-		this.view.copy(target, source);
+		this.host.instance.copy(target, source);
 	}
 
 	generateKey(source) {
-		return this.view.generateKey(source);
+		return this.host.instance.generateKey(source);
 	}
 
 	extract(key, type): ColumnModel {
-		return this.view.extract(key, type);
+		return this.host.instance.extract(key, type);
 	}
 
 	register(column: ColumnModel) {
-		this.view.register(column);
+		this.host.instance.register(column);
 	}
 
 	delete(key: string) {
-		this.view.delete(key);
-	}
-
-	get view() {
-		if (this.host) {
-			return this.host;
-		}
-
-		const canCopy = (key: string, source, target) =>
-			target.hasOwnProperty(key) && !isUndefined(source[key]);
-
-		this.host = new ColumnListHost(this.plugin.model, canCopy, parseFactory);
-
-		return this.host;
+		this.host.instance.delete(key);
 	}
 }