+++ projects/ngx-qgrid/src/lib/table/table-core.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
+import { ColumnModelPin } from '@qgrid/core/column-type/column.model';
 
 @Injectable()
 export class TableCoreService {
-	pin: string;
+	pin: ColumnModelPin;
 }