+++ projects/qgrid-core/foot/foot.let.d.ts
@@ -1,4 +1,4 @@
-import { ColumnModel } from '../column-type/column.model';
+import { ColumnModel, ColumnModelPin } from '../column-type/column.model';
 import { Model } from '../model/model';
 import { GridPlugin } from '../plugin/grid.plugin';
 import { ColumnView } from '../scene/view/column.view';
@@ -9,7 +9,7 @@ export declare class FootLet {
 	readonly count: number;
 	readonly rows: any[];
 
-	columns(row: any, pin: string): ColumnView[];
+	columns(row: any, pin: ColumnModelPin): ColumnView[];
 	invalidate(model: Model): void;
 	value(column: ColumnModel): any;
 	valueFactory: (column: ColumnModel) => (row: any, value?: any) => any;