+++ projects/qgrid-core/scene/render/render.d.ts
@@ -1,4 +1,4 @@
-import { ColumnModel } from '../../column-type/column.model';
+import { ColumnModel, ColumnModelPin } from '../../column-type/column.model';
 import { ColumnView } from '../view/column.view';
 import { RenderStrategy } from './render.strategy';
 import { GridPlugin } from '../../plugin/grid.plugin';
@@ -8,7 +8,7 @@ export declare class Renderer {
 
 	defaultStrategy: RenderStrategy;
 
-	columns(row: any, pin: string, rowIndex: number): ColumnView[];
+	columns(row: any, pin: ColumnModelPin, rowIndex: number): ColumnView[];
 	rowspan(row: any, column: ColumnView, rowIndex: number, columnIndex: number): number;
 	colspan(row: any, column: ColumnView, rowIndex: number, columnIndex: number): number;
 
@@ -18,5 +18,5 @@ export declare class Renderer {
 	getLabel(row: any, column: ColumnModel, rowIndex: number, columnIndex: number): any;
 	setLabel(row: any, column: ColumnModel, value: any, rowIndex: number, columnIndex: number): void;
 
-	readonly rows: { left: any[], right: any[], null: any[] };
+	readonly rows: { left: any[], right: any[], mid: any[] };
 }