+++ projects/qgrid-core/scene/scene.state.d.ts
@@ -1,6 +1,11 @@
 import { ColumnView } from './view/column.view';
 import { RowView } from './view/row.view';
 
+export declare type SceneStateStatus = 'idle' | 'start' | 'pull' | 'push' | 'stop';
+export declare type SceneStateColumnArea =  { left: ColumnView[], right: ColumnView[], mid: ColumnView[] };
+export declare type SceneStateColumnLine = ColumnView[];
+export declare type SceneStateColumnRows = ColumnView[][];
+
 /**
  * A class that contains results of q-grid invalidate.
  */
@@ -14,7 +19,7 @@ export declare class SceneState {
 	 * * `push` request UI to draw a model.
 	 * * `stop` scene in the stable state.
 	 */
-	status: string;
+	status: SceneStateStatus;
 
 	/**
 	 * List of rows to render.
@@ -25,8 +30,8 @@ export declare class SceneState {
 	 * Column rendering object.
 	 */
 	column: {
-		rows: ColumnView[][],
-		line: ColumnView[],
-		area: {}
+		rows: SceneStateColumnRows,
+		line: SceneStateColumnLine,
+		area: SceneStateColumnArea
 	};
 }