+++ projects/qgrid-core/scene/scene.state.js
@@ -1,13 +1,13 @@
 export class SceneState {
 	constructor() {
-		this.status = 'stop'; // start | pull | push | stop
+		this.status = 'idle'; // idle | start | pull | push | stop
 		this.rows = [];
 		this.column = {
 			rows: [],
 			line: [],
 			area: {
 				left: [],
-				null: [],
+				mid: [],
 				right: []
 			}
 		};